services/mcp-server/pyproject.toml
[project]
name = "infratools-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp[server]",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic>=2",
    "azure-search-documents",
    "azure-identity",
    "jinja2",
    "ddtrace",
    "datadog-api-client",
    "openai>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
